+++ RDBqc.Rcheck/00_pkg_src/RDBqc/R/MEDBS_ML_check.R
@@ -31,25 +31,26 @@ MEDBS_ML_check <- function(data, SP, MS, GSA,verbose=TRUE) {
   ML_tab[ML_tab$SEX == FALSE, "SEX"] <- "F"
   Summary_ML_tab <- LENGTHCLASS <- PRM <- COUNTRY <- YEAR <- START_YEAR <- END_YEAR <- SPECIES <- SEX <- NULL
   Summary_ML_tab <- aggregate(ML_tab$SEX, by = list(ML_tab$COUNTRY, ML_tab$AREA, ML_tab$START_YEAR, ML_tab$END_YEAR, ML_tab$SPECIES, ML_tab$SEX), FUN = "length")
-  colnames(Summary_ML_tab) <- c("COUNTRY", "YEAR", "START_YEAR", "END_YEAR", "SPECIES", "SEX")
+  colnames(Summary_ML_tab) <- c("COUNTRY", "YEAR", "START_YEAR", "END_YEAR", "SPECIES", "SEX", "COUNT")
 
+  Summary_table_ML <- Summary_ML_tab
   Summary_ML_tab <- Summary_ML_tab[1:nrow(Summary_ML_tab), 1:(ncol(Summary_ML_tab) - 1)]
 
   output <- list()
   l <- length(output)+1
-  output[[l]] <- Summary_ML_tab
+  output[[l]] <- Summary_table_ML
   names(output)[[l]] <- "summary table"
 
-  p <- ggplot(data = ML_tab, aes(x = LENGTHCLASS, y = PRM, col = factor(START_YEAR))) +
-    geom_line(stat = "identity") +
-    facet_grid(AREA + COUNTRY ~ SEX) +
-    labs(color='Years') +
-      ggtitle(SP)
-  print(p)
-
-  l <- length(output)+1
-  output[[l]] <- p
-  names(output)[[l]] <- paste("ML",SP,MS,GSA,sep=" _ ")
+  # p <- ggplot(data = ML_tab, aes(x = LENGTHCLASS, y = PRM, col = factor(START_YEAR))) +
+  #   geom_line(stat = "identity") +
+  #   facet_grid(AREA + COUNTRY ~ SEX) +
+  #   labs(color='Years') +
+  #     ggtitle(SP)
+  # print(p)
+  #
+  # l <- length(output)+1
+  # output[[l]] <- p
+  # names(output)[[l]] <- paste("ML",SP,MS,GSA,sep=" _ ")
 
   for (i in unique(as.character(ML_tab$SEX))) {
     p <- ggplot(data = ML_tab[ML_tab$SEX %in% i, ], aes(x = LENGTHCLASS, y = PRM, col = factor(START_YEAR))) +