+++ RDBqc.Rcheck/RDBqc/NEWS.md
@@ -1,6 +1,7 @@
 - RDBqc v0.0.10
   - update of MEDBS functions
   - new functions to check SOP and ALK
+  - implemented the treatment of NA values for VESSEL_LENGTH, GEAR, FISHERY and MESH_SIZE_RANGE in MEDBS functions 
 - RDBqc v0.0.9
   - toupper in colnames of MEDBS checks